+++ CHANGES.md
@@ -1,10 +1,16 @@
-## Version 0.4.8?
+## Version 0.4.8
+
+- Removed framework and init modules, and consolidated in sim module (so just need: from netpyne import sim)
 
 - Added option to specify weight scale factor separately for each cell model (issue #69)
 
 - Conn rules can now have list of synMechs (eg. [AMPA, NMDA]) and synMechWeightFactor (eg. [1.0, 0.1]) (issue #69)
 
+- Conn rules can provide list of weights and/or delays for each synMech in list (issue #69)
+
 - Conn rules allow synsPerConn to be described functionally (eg. 'uniform(5,2)') (issue #69)
+
+- Added separate wrapper sim functions to create and simulate the network  
 
 - Fixed bug so can create Python and NEURON objects for connections independently (issue #69)